doc-summarizer-be/pyproject.toml
[project]
name = "doc-summarizer-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Gokul Sudheesh",email = "gokulsudheesh1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.65,<0.4.0)",
    "langchain-nvidia-ai-endpoints (>=0.3.10,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)"
]

[tool.poetry]
packages = [{include = "doc_summarizer_be", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
